refactor: util 패키지 삭제
